Refactor task pane state management into a dedicated hook
